% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate_properties.R
\name{create_plate_properties}
\alias{create_plate_properties}
\title{Create plate properties.}
\usage{
create_plate_properties(
  row_range,
  column_range,
  well_contents,
  max_conc,
  direction = "horizontal",
  dil_factor = 0.5,
  constant_concentration = FALSE
)
}
\arguments{
\item{row_range}{Vector of letters to generate values for (e.g LETTERS[1:3])}

\item{column_range}{Vector of numbers to generate values for (e.g. 1:3)}

\item{well_contents}{String for label of what is in the well (e.g. "Sample1")}

\item{max_conc}{Maximum concentration of dilution series for values.}

\item{direction}{String either "horizontal" (default) or "vertical". "horizontal" makes dilution series go across numbered colums, "vertical" makes diultion series go down the letterd rows.}

\item{dil_factor}{Dilution factor for generating dilution series. Default = 0.5.}

\item{constant_concentration}{TRUE/FALSE. TRUE = only 1 concentration for all values. FALSE = generating dilution series.}
}
\description{
Generates concentrations, row / column labels and well contents from input.
}
\examples{

create_plate_properties(LETTERS[1:3], 1:3, 500, "Normal")
create_plate_properties(LETTERS[1:3], 1:3, 500, "vertical_values", direction = "vertical")
create_plate_properties(LETTERS[1:3], 1:3, 500, "constant_values", constant_concentration = T)
}
\keyword{concentrations}
\keyword{dilution,}
\keyword{plate,}
