% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plate.R
\name{make_plate}
\alias{make_plate}
\title{Make a plate-properties dataframe.}
\usage{
make_plate(rows, cols, sample, conc = NA, dil = NA, direction = "horizontal")
}
\arguments{
\item{rows}{Vector of letters or numbers to generate row labels. If numeric LETTERS[rows]}

\item{cols}{Vector of numbers to generate column values.}

\item{conc}{If no dilution, constant concentration. If dilution, highest concentration in the dilution series.}

\item{dil}{Numeric between 0 and 1 to represent dilution series. 1:1 dilution = 0.5.}

\item{direction}{If !is.na(dil) then direction dictates if this dilution is performed across the columns (horizontal) or down the rows (vertical). Defaults to horizontal.}
}
\description{
Function for creating plate properties given column, row range and sample contents.
Optional concentration, dilution factor and direction (for dilution series).
}
\examples{

make_plate(rows = 1:3, cols = 1:2, sample = c("foo", "bar"))
make_plate(rows = 1:3, cols = 1:2, sample = c("foo", "bar"), conc = 500, dil = 0.5)
}
\keyword{concentrations}
\keyword{dilution,}
\keyword{plate,}
