% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wells.R
\name{make_wells}
\alias{make_wells}
\title{Create well labels.}
\usage{
make_wells(let, num)
}
\arguments{
\item{let}{Character or numeric vector for letters component. If numeric will become LETTERS[let]}

\item{num}{Numeric vector for column numbers.}
}
\description{
Generates well labels (e.g. A1-H12) for given range of letters and numbers.
Letters can either be a character of numeric range.
}
\examples{

make_wells(LETTERS[1:2], 1:8)
make_wells(1:2, 1:8)
}
\keyword{plate,}
\keyword{wells}
