% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowfill_conc.R
\name{rowfill_conc}
\alias{rowfill_conc}
\title{Generate dilution series.}
\usage{
rowfill_conc(max_conc, num_dil, dil_factor = 0.5)
}
\arguments{
\item{max_conc}{The maximum value in the dilution series.}

\item{num_dil}{The number of values in the dilution series (indluding the max_conc)}

\item{dil_factor}{The dilution factor for the series. Defatuls = 0.5}
}
\description{
Generates a dilution series starting at max_conc. Default diltuion factor = 0.5.
}
\examples{

rowfill_conc(100, 5, dil_factor = 0.5)
}
\keyword{concentration}
\keyword{dilution,}
\keyword{plate,}
